backend/requirements.txt
fastapi
uvicorn
supabase
groq
pydantic
python-dotenv